Remove tracked man/multigifview.1 file
Modifying the man file in the release action messes up versioning which
requires the repo state to be clean. To get around this, remove the
tracked file. An example stub man-file is included with a different
name, and a note added to DEVELOPMENT.md on how to install from the git
repo if necessary.